+++ andreas-technology-v3/src/app/layout.tsx
@@ -33,6 +33,7 @@ export const metadata: Metadata = {
     },
     icons: {
         icon: [
+            { url: '/favicon.ico', sizes: '32x32', type: 'image/png' },
             { url: '/favicons/favicon-32x32.png', sizes: '32x32', type: 'image/png' },
             { url: '/favicons/favicon-16x16.png', sizes: '16x16', type: 'image/png' },
         ],