+++ andreas-technology-v3/src/app/robots.ts
@@ -4,7 +4,11 @@ export const dynamic = 'force-static'
 
 export default function robots(): MetadataRoute.Robots {
   return {
-    rules: { userAgent: '*', allow: '/' },
+    rules: {
+      userAgent: '*',
+      allow: '/',
+      disallow: ['/opengraph-image'],
+    },
     sitemap: 'https://andreas.technology/sitemap.xml',
   }
 }