+++ andreas-technology-v3/src/app/sitemap.ts
@@ -10,11 +10,5 @@ export default function sitemap(): MetadataRoute.Sitemap {
             changeFrequency: 'monthly',
             priority: 1,
         },
-        {
-            url: 'https://andreas.technology/thesis-presentation.html',
-            lastModified: new Date(),
-            changeFrequency: 'monthly',
-            priority: 0.8,
-        },
     ]
 }